qubes: add fedora version to template naming convention
When upgrading to a new Fedora template it will be useful
to leave the old template until migration is complete. The
version name is appended to the end of the template name.
Installation of i2pd to the newly cloned template for
i2nix-gateway should be the only required change. Ensure
i2pd.service is enabled.